Built-in lexicons

A list containing all built-in lexicons as a data.table with two columns: a x column with the words, and a y column with the polarities. The list element names incorporate consecutively the name and language (based on the two-letter ISO code convention as in stopwords), and "_tr" as suffix if the lexicon is translated. The translation was done via Microsoft Translator through Microsoft Word. Only the entries that conform to the original language entry after retranslation, and those that have actually been translated, are kept. The last condition is assumed to be fulfilled when the translation differs from the original entry. All words are unigrams and in lowercase. The built-in lexicons are the following:

FEEL_en_tr

FEEL_fr (Abdaoui, Azé, Bringay and Poncelet, 2017)

FEEL_nl_tr

GI_en (General Inquirer, i.e. Harvard IV-4 combined with Laswell)

GI_fr_tr

GI_nl_tr

HENRY_en (Henry, 2008)

HENRY_fr_tr

HENRY_nl_tr

LM_en (Loughran and McDonald, 2011)

LM_fr_tr

LM_nl_tr

Other useful lexicons can be found in the lexicon package, more specifically the datasets preceded by hash_sentiment_.

data("list_lexicons")

Format

A list with all built-in lexicons, appropriately named as "NAME_language(_tr)".

Source

FEEL lexicon. Retrieved November 1, 2017.

GI lexicon. Retrieved November 1, 2017.

HENRY lexicon. Retrieved November 1, 2017.

LM lexicon. Retrieved November 1, 2017.

References

Abdaoui, Azé, Bringay and Poncelet (2017). FEEL: French Expanded Emotion Lexicon. Language Resources & Evaluation 51, 833-855, doi:10.1007/s10579-016-9364-5 .

Henry (2008). Are investors influenced by how earnings press releases are written?. Journal of Business Communication 45, 363-407, doi:10.1177/0021943608319388 .

Loughran and McDonald (2011). When is a liability not a liability? Textual analysis, dictionaries, and 10-Ks. Journal of Finance 66, 35-65, doi:10.1111/j.1540-6261.2010.01625.x .

Examples

data("list_lexicons", package = "sentometrics")
list_lexicons[c("FEEL_en_tr", "LM_en")]
#> $FEEL_en_tr
#> Key: <x>
#>                  x     y
#>             <char> <num>
#>    1:  abandonment    -1
#>    2: abbreviation     1
#>    3:     abdicate    -1
#>    4:    abdominal     1
#>    5:     abductor    -1
#>   ---                   
#> 4411:         zeal     1
#> 4412:      zealous    -1
#> 4413:         zero    -1
#> 4414:       zipper     1
#> 4415:      zoology     1
#> 
#> $LM_en
#> Key: <x>
#>                  x     y
#>             <char> <num>
#>    1:      abandon    -1
#>    2:    abandoned    -1
#>    3:   abandoning    -1
#>    4:  abandonment    -1
#>    5: abandonments    -1
#>   ---                   
#> 2705:   wrongdoing    -1
#> 2706:  wrongdoings    -1
#> 2707:     wrongful    -1
#> 2708:   wrongfully    -1
#> 2709:      wrongly    -1
#>