sentometrics: An Integrated Framework for Textual Sentiment Time Series Aggregation and Prediction

The sentometrics package is an integrated framework for textual sentiment time series aggregation and prediction. It accounts for the intrinsic challenge that, for a given text, sentiment can be computed in many different ways, as well as the large number of possibilities to pool sentiment across texts and time. This additional layer of manipulation does not exist in standard text mining and time series analysis packages. The package therefore integrates the fast quantification of sentiment from texts, the aggregation into different sentiment time series and the optimized prediction based on these measures.

Note

Please cite the package in publications. Use citation("sentometrics").

Main functions

Corpus (features) generation: sento_corpus, add_features, as.sento_corpus

Sentiment computation and aggregation into sentiment measures: ctr_agg, sento_lexicons, compute_sentiment, aggregate.sentiment, as.sentiment, sento_measures, peakdocs, peakdates, aggregate.sento_measures

Sparse modeling: ctr_model, sento_model

Prediction and post-modeling analysis: predict.sento_model, attributions

References

Ardia, Bluteau and Boudt (2019). Questioning the news about economic growth: Sparse forecasting using thousands of news-based sentiment values. International Journal of Forecasting 35, 1370-1386, doi: 10.1016/j.ijforecast.2018.10.010 .

Ardia, Bluteau, Borms and Boudt (2021). The R package sentometrics to compute, aggregate and predict with textual sentiment. Journal of Statistical Software 99(2), 1-40, doi: 10.18637/jss.v099.i02 .

See also

Author

Maintainer: Samuel Borms borms_sam@hotmail.com (ORCID)

Authors:

David Ardia david.ardia@hec.ca (ORCID)

Keven Bluteau keven.bluteau@unine.ch (ORCID)

Kris Boudt kris.boudt@vub.be (ORCID)

Other contributors:

Jeroen Van Pelt jeroenvanpelt@hotmail.com [contributor]

Andres Algaba andres.algaba@vub.be [contributor]