Texts (not) relevant to the U.S. economy

A collection of texts annotated by humans in terms of relevance to the U.S. economy or not. The texts come from two major journals in the U.S. (The Wall Street Journal and The Washington Post) and cover 4145 documents between 1995 and 2014. It contains following information:

id. A character ID identifier.

date. Date as "yyyy-mm-dd".

texts. Texts in character format.

wsj. Equals 1 if the article comes from The Wall Street Journal.

wapo. Equals 1 if the article comes from The Washington Post (complementary to `wsj').

economy. Equals 1 if the article is relevant to the U.S. economy.

noneconomy. Equals 1 if the article is not relevant to the U.S. economy (complementary to `economy').

data("usnews")

Format

A data.frame, formatted as required to be an input for sento_corpus.

Source

Economic News Article Tone and Relevance. Retrieved November 1, 2017.

Examples

data("usnews", package = "sentometrics")
usnews[3192, "texts"]

#> [1] "Author Edward P. Lazear As Washington debates the fate of the 2001 and 2003 tax cuts, many lawmakers have fallen into a logical trap of their own making. Although they recognize that tax increases hurt the economy, they argue that our huge deficit requires Congress to raise revenue through a tax hike. This argument rests on the flawed premise that we can reduce the deficit only by increasing taxes, as if high levels of spending are a given. Not so. To reduce spending and reignite growth, this Congress or its successor should take two actions. First, immediately cut the level of spending that has been increased so dramatically since 2008. Second, institute an inflation-minus-one rule to constrain future spending increases. Much public discussion focuses on the deficit, which is indeed at critical levels of around 10 of GDP. But even if President Obama succeeds at lowering the deficit to 4 of GDP by 2013, our public-debt-to-GDP ratio will still be dangerously high, at over 70, or nearly twice what it was during the Bush years. As the economists Carmen Reinhart and Kenneth Rogoff have shown in the journal American Economic Review, such high debt-to-GDP ratios are associated with low growth."

usnews[1:5, c("id", "date", "texts")]

#>          id       date
#> 1 830981846 1995-01-02
#> 2 842617067 1995-01-05
#> 3 830982165 1995-01-05
#> 4 830982389 1995-01-08
#> 5 842615996 1995-01-09
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         texts
#> 1                                                                                                                                                                                                                                                                     In 1994, the good times for Washington area stocks were short-lived. They lasted only until Feb. 4, the day that the Federal Reserve began to raise interest rates. After that, and for the next 11 months, it was all downhill. By the time the turbulent year was over, almost two-thirds of Washington area stocks had lost money. The losses cut across a wide band and hurt financial, retail, biotech, telecommunications and transportation stocks. It also was the kind of year in which the relatively few winners fell into two categories companies with products or services that were in high demand and companies that, having faltered financially, were in a turnaround mode. As often happens in periods of market turmoil, the big blue-chip stocks listed in the Dow Jones industrial average and the Standard & Poor's 500 stock index fared somewhat better than the stocks of the hundreds of smaller companies in a business community that extends from Baltimore to Richmond. It was a bear market and if you look at the Dow and the S&P, it doesn't reflect what happened in the broad market, said Prabha S. Carpenter, portfolio manager of the Growth Fund of Washington, a 33 million fund that invests in three dozen local stocks. Carpenter's fund was down 9.3 percent for the year.
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             NEW YORK -- Small stocks rose in light trading as investors' tentativeness continued to prevent the market from establishing direction. Some selective bargain-hunting took place, but investors mostly reacted to news on individual issues, leaving a scattering of gains and declines. Major indexes traded in very narrow ranges. The Russell 2000 Index, which measures small-capitalization issues that trade on the Nasdaq Stock Market and the other major exchanges, inched up 0.41, or 0.17, to 247.65. The Nasdaq Composite gained 2.26, or 0.30, to finish at 745.84. The breadth of the market was positive for the day, with advancers ahead of decliners, 1,752 to 1,424, on Nasdaq total volume of 290.3 million shares, compared with the previous day's 248.7 million shares.
#> 3                                                                                                                                                                                   The Dow Jones industrial average climbed 19.17 points to close at 3857.65, and made most of its gain in the final 30 minutes. Advancing issues outnumbered declining ones by about 3 to 2 on the New York Stock Exchange, where trading volume rose to The bond market rally that lifted stocks was triggered by revived strength in the dollar and plunging precious metals prices. The price of the Treasurys main 30-year bond advanced 2332 point, or 7.19 per 1,000 in face value. The yield fell to 7.85 percent from 7.91 percent. Traders and investors bought bonds amid a rally in the dollar to a 21-week high against the Japanese yen. Traders believed the stronger dollar could draw some foreign investors to dollar-denominated investments, which include U.S. Treasury securities. In addition, the rising U.S. currency reassured investors worried about inflation, which tends to diminish the value of fixed-income securities. 100.68late Tuesday. The dollar also changed hands in New York at 1.5593 German marks, up from 1.5562. By midday Thursday in Tokyo, the dollar was slightly lower, trading at 101.10 yen. 4.7percent against the dollar, reflecting continued weakness in Mexicos financial markets and disenchantment among many investors with its plan to bolster its troubled economy. Late in Mexico City, the dollar fetched 5.5750 pesos vs. 5.3250 Tuesday.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           First Months Performance Has Predicted the Markets Direction Accurately Many Times Before. But Will It Do So Again This Year First Months Performance Has Predicted the Markets Direction Accurately Many Times Before. But Will It Do So Again This Year mhe overall trend of the market for the past three years has been set during January. So far in 1995, January again has been a good month. NEW YORKIts nail biting time on Wall Street as the stock markets performance this month provides the January barometer that is supposed to predict whether the year on Wall Street will be good or bad. Yale Hirsch, publisher of Stock Traders Almanac, said that the month of January has proved to be an incredible forecaster of things to come for the rest of the year. Nothing beats the January Barometer, he said. Since 1950, no other indicator has predicted the annual course of the market with such accuracy.
#> 5 PITTSBURGH -- PNC Bank Corp., moving to reduce its vulnerability to rising interest rates, said it sold 1.8 billion in securities last month as part of a plan to shrink its investment portfolio by 5.9 billion by the end of 1995. The bank-holding company said it will incur a 79 million loss for the fourth quarter from the recent sale of fixed-rate securities. It also said it won't replace a further 4.1 billion in securities that mature this year, reducing its total portfolio to about 17 billion by the end of 1995 from 22.9 billion at the end of the 1994 third quarter. The move will reduce PNC's total assets by 9.4, to about 58 billion. PNC indicated last fall that it would restructure its investment portfolio after steep interest-rate increases produced heavy securities losses. At the time, PNC said it expected those losses would cut 1994 fourth-quarter earnings by 7 and 1995 earnings by about 15. PNC isn't alone in selling securities because of rising interest rates. I assume virtually everybody is, said Dennis Shea, an analyst at Morgan Stanley & Co. Most banks are, I would believe, net sellers of securities right now. They're trying to reduce their liability-sensitive positions. To further reduce vulnerability, PNC recently bought interest-rate swaps with fixed payments and interest-rate caps. The new swaps and caps will cost PNC about 70 million in 1995 if rates stay the same, said Thomas H. O'Brien, chairman, chief executive and president. If rates keep rising, the cost of the swaps and caps will drop.